Crate tokio_io_timeout

Tokio wrappers which apply timeouts to IO operations.

These timeouts are analogous to the read and write timeouts on traditional blocking sockets. A timeout countdown is initiated when a read/write operation returns Poll::Pending. If a read/write does not return successfully before the countdown expires, an io::Error with a kind of TimedOut is returned.

Structs

TimeoutReader

An AsyncReader which applies a timeout to read operations.

TimeoutStream

A stream which applies read and write timeouts to an inner stream.

TimeoutWriter

An AsyncWriteer which applies a timeout to write operations.