Struct tokio_io_timeout::TimeoutStream

pub struct TimeoutStream<S> { /* private fields */ }

A stream which applies read and write timeouts to an inner stream.

Implementations

impl<S> TimeoutStream<S>
where S: AsyncRead + AsyncWrite,

pub fn new(stream: S) -> TimeoutStream<S>

Returns a new TimeoutStream wrapping the specified stream.

There is initially no read or write timeout.

pub fn read_timeout(&self) -> Option<Duration>

Returns the current read timeout.

pub fn set_read_timeout(&mut self, timeout: Option<Duration>)

Sets the read timeout.

This can only be used before the stream is pinned; use set_read_timeout_pinned otherwise.

pub fn set_read_timeout_pinned(self: Pin<&mut Self>, timeout: Option<Duration>)

Sets the read timeout.

This will reset any pending read timeout. Use set_read_timeout instead if the stream has not yet been pinned.

pub fn write_timeout(&self) -> Option<Duration>

Returns the current write timeout.

pub fn set_write_timeout(&mut self, timeout: Option<Duration>)

Sets the write timeout.

This can only be used before the stream is pinned; use set_write_timeout_pinned otherwise.

pub fn set_write_timeout_pinned(self: Pin<&mut Self>, timeout: Option<Duration>)

Sets the write timeout.

This will reset any pending write timeout. Use set_write_timeout instead if the stream has not yet been pinned.

pub fn get_ref(&self) -> &S

Returns a shared reference to the inner stream.

pub fn get_mut(&mut self) -> &mut S

Returns a mutable reference to the inner stream.

pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut S>

Returns a pinned mutable reference to the inner stream.

pub fn into_inner(self) -> S

Consumes the stream, returning the inner stream.